requirements.txt
ultralytics
dill
cvzone
opencv-python
pillow
numpy
ipython
joblib
tensorflow
scikit-image